More or less a total rewrite, added documentation and tests
Relized that I couldn't support dot syntax or use Scopes as proxy
objects in lexically flattened dictionaries. The issue comes down to one
of not being able to differentiate values from structure. If you
previously used a dictionary that was defined in YAML as a value, you
would be disappointed that it was actually a Scope (proxy object)!

I decided that the usefulness of the proxy objects was:
* dox syntax when accessing "variables"
* supporting "this" and "up" special variables to traverse the hierarchy
and that they weren't useful enough to bother keeping.

Also condensed everythign into a single < 300 line module.
Fix several issues including:
* inability to evaluate Python in templates like with f-strings
* import's poor syntax and support for importing nested objects
* improved the stacktrace